refactored: remote config and default value of home screen preview anime
