+++ core/domain/src/main/java/dev/enesky/core/domain/manager/remoteconfig/FetchStatus.kt
@@ -1,0 +1,11 @@
+package dev.enesky.core.domain.manager.remoteconfig
+
+/**
+ * Created by Enes Kamil YILMAZ on 07/03/2024
+ */
+
+enum class FetchStatus {
+    LOADING,
+    SUCCESS,
+    ERROR,
+}
