+++ core/domain/src/main/java/dev/enesky/core/domain/manager/remoteconfig/RemoteConfigArgs.kt
@@ -1,0 +1,18 @@
+package dev.enesky.core.domain.manager.remoteconfig
+
+/**
+ * Created by Enes Kamil YILMAZ on 07/03/2024
+ */
+
+object RemoteConfigArgs {
+
+    object Keys {
+        const val HOME_SCREEN_PREVIEW_ANIME_ID = "home_screen_preview_anime_id"
+    }
+
+    object DefaultValues {
+        const val ONE_PIECE_ANIME_ID = "21"
+        const val VINLAND_SAGA_SEASON_2_ANIME_ID = "49387"
+    }
+
+}
