+++ core/domain/src/main/java/dev/enesky/core/domain/manager/remoteconfig/RemoteConfigManager.kt
@@ -1,10 +1,10 @@
-package dev.enesky.core.domain.manager
+package dev.enesky.core.domain.manager.remoteconfig
 
 import com.google.firebase.ktx.Firebase
 import com.google.firebase.remoteconfig.ktx.remoteConfig
 import com.google.firebase.remoteconfig.ktx.remoteConfigSettings
 import dev.enesky.core.common.BuildConfig
-import dev.enesky.core.domain.manager.RemoteConfigKeys.HOME_SCREEN_PREVIEW_ANIME_ID
+import dev.enesky.core.domain.manager.remoteconfig.RemoteConfigArgs.Keys.HOME_SCREEN_PREVIEW_ANIME_ID
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 
@@ -14,32 +14,35 @@ import kotlinx.coroutines.flow.StateFlow
 
 object RemoteConfigManager {
 
-    private val _configStatus = MutableStateFlow(FetchStatus.LOADING)
     val configStatus: StateFlow<FetchStatus>
         get() = _configStatus
+    private val _configStatus = MutableStateFlow(FetchStatus.LOADING)
+    private val remoteConfig by lazy { Firebase.remoteConfig }
+    private const val FETCH_INTERVAL_DEBUG: Long = 10 // Fetch every 10 seconds in debug
+    private const val FETCH_INTERVAL: Long = 60 * 60 * 1 // Fetch every 1 hour
+    private const val FETCH_RETRY: Long = 30
 
-    private val remoteConfig by lazy {
-        Firebase.remoteConfig
+    object Values {
+        val homeScreenPreviewAnimeId: Int
+            get() = remoteConfig.getLong(HOME_SCREEN_PREVIEW_ANIME_ID).toInt()
     }
 
+    /**
+     * Initialize Remote Config with default values and fetch the remote config
+     */
     suspend fun init() {
-        val fetchInterval: Long = if (BuildConfig.DEBUG) {
-            10 // Fetch every 10 seconds in debug
-        } else {
-            60 * 60 * 1 // Fetch every 1 hour in production
-        }
-        val fetchRetry = 30L
+        val fetchInterval: Long = if (BuildConfig.DEBUG) FETCH_INTERVAL_DEBUG else FETCH_INTERVAL
 
         remoteConfig.apply {
             setConfigSettingsAsync(
                 remoteConfigSettings {
                     minimumFetchIntervalInSeconds = fetchInterval
-                    fetchTimeoutInSeconds = fetchRetry
+                    fetchTimeoutInSeconds = FETCH_RETRY
                 },
             )
             setDefaultsAsync(
                 mutableMapOf<String, Any>(
-                    HOME_SCREEN_PREVIEW_ANIME_ID to "10",
+                    HOME_SCREEN_PREVIEW_ANIME_ID to RemoteConfigArgs.DefaultValues.VINLAND_SAGA_SEASON_2_ANIME_ID,
                 ),
             )
         }
@@ -60,17 +63,4 @@ object RemoteConfigManager {
             }
         }
     }
-
-    val homeScreenPreviewAnimeId: Int
-        get() = remoteConfig.getLong(HOME_SCREEN_PREVIEW_ANIME_ID).toInt()
-}
-
-object RemoteConfigKeys {
-    const val HOME_SCREEN_PREVIEW_ANIME_ID = "home_screen_preview_anime_id"
-}
-
-enum class FetchStatus {
-    LOADING,
-    SUCCESS,
-    ERROR,
 }
