+++ feature/home/src/main/java/dev/enesky/feature/home/HomeViewModel.kt
@@ -10,7 +10,7 @@ import dev.enesky.core.common.delegate.UiStateDelegate
 import dev.enesky.core.common.enums.AnimeFilter
 import dev.enesky.core.common.result.Result
 import dev.enesky.core.common.result.asResult
-import dev.enesky.core.domain.manager.RemoteConfigManager
+import dev.enesky.core.domain.manager.remoteconfig.RemoteConfigManager
 import dev.enesky.core.domain.usecase.DetailedAnimeUseCase
 import dev.enesky.core.domain.usecase.TopAnimePagingUseCase
 import dev.enesky.feature.home.helpers.HomeEvents
